Create a mock file for config.schema.ts
